ml_project/requirements/requirements.txt
hydra-core==1.0.6
omegaconf==2.0.6
pandas==1.2.1
pandas-profiling==2.12.0
PyYAML==5.4.1
scikit-learn==0.23.2
numpy==1.19.2
pytest==6.1.1